Refactor code to be more extensible
Controls of the players are no longer hard coded and can be adjusted by keybode in the inspector.
Players are also no longer using different tags and in different layers making it easier to extend to multiplayer mode.
Player number is also derived from the number of existing players making it possible to have more than 2 players.